ginny/manage-users (#16)
* feat: create getUser usecase

* refactor: use getUser usecase to retrieve user when get and update event

* feat: add roles to user and permissions

* feat: create listing events by users and by time

+++ usecase/events/deleteEvent.js
@@ -1,11 +1,25 @@
 const eventRepo = require('../../repositories/events');
 const createError = require('http-errors');
 
-const handle = async (eventId) => {
+const handle = async (eventId, currentUser) => {
+  // Find the event to delete
   const foundEvent = await eventRepo.getById(eventId);
+
   if (foundEvent) {
-    await eventRepo.remove(foundEvent._id);
-    return foundEvent;
+    // Check permission; ONLY admin and owner can delete event
+    const adminListArr = process.env.ADMINLIST.split(',');
+    if (
+      adminListArr.includes(currentUser) ||
+      foundEvent.createdBy === currentUser
+    ) {
+      await eventRepo.remove(foundEvent._id);
+      return foundEvent;
+    } else {
+      throw createError(
+        400,
+        'Your account is not allowed to delete the following event'
+      );
+    }
   } else {
     throw createError(404, 'No event with that id found!');
   }
